+++ CHANGELOG.md
@@ -57,6 +57,7 @@
   interfaces. Specifying interfaces that allow forwarding requests to MMDS is done
   by adding the network interface's ID to the `network_interfaces` field of PUT
   `/mmds/config` request's body.
+- Configuring `ht_enabled: true` on aarch64 via the API is forbidden.
 
 ### Fixed